# openapi.yaml
openapi: 3.1.0
info:
  title: Vessica Knowledge API
  version: 1.3.0
  description: |
    Workspace-scoped API for authoritative artifacts, durable memories, entities,
    relationships, ranked context, workflow episodes, and verified snapshots.
    All responses use the `vessica.knowledge/v1` envelope. All writes require
    `Idempotency-Key`, `X-Actor-ID`, `X-Actor-Type`, and
    `X-Provenance-Source` headers.
  license: {name: Apache-2.0, identifier: Apache-2.0}
servers:
  - url: https://{host}
    variables:
      host: {default: knowledge.example.com}
security:
  - bearerAuth: []
...
paths:
...
  /admin/v1/embeddings/backfill/{job_id}:
    get:
      tags: [Context]
      summary: Read one embedding backfill job
      operationId: getEmbeddingBackfillJob
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Backfill job status, content: {application/json: {schema: {$ref: '#/components/schemas/EmbeddingBackfillEnvelope'}}}}
...
  /v1/status:
    get:
      tags: [Context]
      summary: Read retrieval and index status
      operationId: getKnowledgeStatus
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
      responses:
        '200': {description: Retrieval and embedding status, content: {application/json: {schema: {$ref: '#/components/schemas/StatusEnvelope'}}}}
  /admin/v1/embeddings/backfill:
    post:
      tags: [Context]
      summary: Queue an idempotent embedding backfill for current memories
      operationId: queueEmbeddingBackfill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workspace_id: {type: string}
                mode: {type: string, enum: [missing, all], default: missing}
      responses:
        '202': {description: Backfill queued, content: {application/json: {schema: {$ref: '#/components/schemas/EmbeddingBackfillEnvelope'}}}}
    get:
      tags: [Context]
      summary: Read embedding backfill backlog
      operationId: getEmbeddingBackfill
      parameters: [{$ref: '#/components/parameters/WorkspaceID'}]
      responses:
        '200': {description: Backfill status, content: {application/json: {schema: {$ref: '#/components/schemas/EmbeddingBackfillEnvelope'}}}}
  /v1/search:
    get:
      tags: [Context]
      summary: Search entities, artifacts, and memories
      operationId: searchKnowledge
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
        - {name: q, in: query, schema: {type: string}}
        - {name: object_type, in: query, schema: {type: string, enum: [entity, artifact, memory]}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200': {description: Paginated search results, content: {application/json: {schema: {$ref: '#/components/schemas/SearchPageEnvelope'}}}}
  /v1/context:
    post:
      tags: [Context]
      summary: Assemble ranked context
      operationId: assembleContext
      requestBody: {$ref: '#/components/requestBodies/ContextRequest'}
      responses:
        '200': {description: Ranked context, content: {application/json: {schema: {$ref: '#/components/schemas/ContextEnvelope'}}}}
        '400': {$ref: '#/components/responses/Invalid'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/scopes:
    post:
      tags: [Entities]
      summary: Create a scope
      operationId: createScope
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/ActorID'}, {$ref: '#/components/parameters/ActorType'}, {$ref: '#/components/parameters/ProvenanceSource'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Scope'}}}}
      responses:
        '201': {description: Scope, content: {application/json: {schema: {$ref: '#/components/schemas/ScopeEnvelope'}}}}
        '400': {$ref: '#/components/responses/Invalid'}
  /v1/scopes/{id}:
    get:
      tags: [Entities]
      summary: Get a scope
      operationId: getScope
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}]
      responses:
        '200': {description: Scope, content: {application/json: {schema: {$ref: '#/components/schemas/ScopeEnvelope'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/entities:
    get:
      tags: [Entities]
      summary: List current entities
      operationId: listEntities
      parameters: [{$ref: '#/components/parameters/WorkspaceID'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Paginated entities, content: {application/json: {schema: {$ref: '#/components/schemas/EntityPageEnvelope'}}}}
    post:
      tags: [Entities]
      summary: Create an entity
      operationId: createEntity
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/ActorID'}, {$ref: '#/components/parameters/ActorType'}, {$ref: '#/components/parameters/ProvenanceSource'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Entity'}}}}
      responses:
        '201': {description: Entity, content: {application/json: {schema: {$ref: '#/components/schemas/EntityEnvelope'}}}}
        '400': {$ref: '#/components/responses/Invalid'}
  /v1/entities/{id}:
    get:
      tags: [Entities]
      summary: Get an entity
      operationId: getEntity
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}]
      responses:
        '200': {description: Entity, content: {application/json: {schema: {$ref: '#/components/schemas/EntityEnvelope'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/entities:resolve:
    get:
      tags: [Entities]
      summary: Resolve entities
      operationId: resolveEntities
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
        - {name: q, in: query, required: true, schema: {type: string}}
        - {name: scope, in: query, schema: {type: array, items: {type: string}}, style: form, explode: true}
      responses:
        '200': {description: Matching entities, content: {application/json: {schema: {$ref: '#/components/schemas/EntityListEnvelope'}}}}
  /v1/artifacts:
    get:
      tags: [Artifacts]
      summary: List artifacts
      operationId: listArtifacts
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
        - {name: type, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string, enum: [draft, active, superseded, archived]}}
        - {name: scope, in: query, schema: {type: array, items: {type: string}}, style: form, explode: true}
      responses:
        '200': {description: Artifacts, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactListEnvelope'}}}}
    post:
      tags: [Artifacts]
      summary: Create an artifact
      operationId: createArtifact
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/ActorID'}, {$ref: '#/components/parameters/ActorType'}, {$ref: '#/components/parameters/ProvenanceSource'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}
      responses:
        '201': {description: Artifact, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactEnvelope'}}}}
        '400': {$ref: '#/components/responses/Invalid'}
  /v1/artifacts/{id}:
    get:
      tags: [Artifacts]
      summary: Get an artifact version
      operationId: getArtifact
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}, {$ref: '#/components/parameters/Version'}]
      responses:
        '200': {description: Artifact, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactEnvelope'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/artifacts/{id}/versions:
    get:
      tags: [Artifacts]
      summary: List immutable artifact versions
      operationId: listArtifactVersions
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Artifact versions newest first, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactPageEnvelope'}}}}
...
  /v1/memories:
    get:
      tags: [Memories]
      summary: Search memories
      operationId: searchMemories
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
        - {name: q, in: query, schema: {type: string}}
        - {name: scope, in: query, schema: {type: array, items: {type: string}}, style: form, explode: true}
      responses:
        '200': {description: Memories, content: {application/json: {schema: {$ref: '#/components/schemas/MemoryListEnvelope'}}}}
    post:
      tags: [Memories]
      summary: Create a memory
      operationId: createMemory
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/ActorID'}, {$ref: '#/components/parameters/ActorType'}, {$ref: '#/components/parameters/ProvenanceSource'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Memory'}}}}
      responses:
        '201': {description: Memory, content: {application/json: {schema: {$ref: '#/components/schemas/MemoryEnvelope'}}}}
  /v1/memories:retrieve:
    post:
      tags: [Memories]
      summary: Retrieve authorized current memories with deterministic hybrid ranking
      operationId: retrieveMemories
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MemoryRetrievalRequest'}
      responses:
        '200': {description: Ranked retrieval-v2 memories and diagnostics, content: {application/json: {schema: {$ref: '#/components/schemas/MemoryRetrievalEnvelope'}}}}
        '400': {$ref: '#/components/responses/Invalid'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/memories/{id}:
    get:
      tags: [Memories]
      summary: Get a memory version
      operationId: getMemory
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}, {$ref: '#/components/parameters/Version'}]
      responses:
        '200': {description: Memory, content: {application/json: {schema: {$ref: '#/components/schemas/MemoryEnvelope'}}}}
  /v1/memories/{id}/versions:
    get:
      tags: [Memories]
      summary: List immutable memory versions
      operationId: listMemoryVersions
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/WorkspaceID'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Memory versions newest first, content: {application/json: {schema: {$ref: '#/components/schemas/MemoryPageEnvelope'}}}}
...
  /v1/relationships:
    get:
      tags: [Relationships]
      summary: List incoming and outgoing relationships for an object
      operationId: listRelationships
      parameters:
        - {$ref: '#/components/parameters/WorkspaceID'}
        - {name: object_id, in: query, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200': {description: Paginated relationships, content: {application/json: {schema: {$ref: '#/components/schemas/RelationshipPageEnvelope'}}}}
    post:
      tags: [Relationships]
      summary: Create a relationship assertion
      operationId: createRelationship
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/ActorID'}, {$ref: '#/components/parameters/ActorType'}, {$ref: '#/components/parameters/ProvenanceSource'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Relationship'}}}}
      responses:
        '201': {description: Immutable relationship assertion, content: {application/json: {schema: {$ref: '#/components/schemas/RelationshipEnvelope'}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: Workspace-scoped ordinary API token}
    exportBearer: {type: http, scheme: bearer, description: Separately scoped export/import token}
  parameters:
...
    WorkspaceID: {name: workspace_id, in: query, required: false, schema: {type: string}}
...
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}}
    Limit: {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
...
  schemas:
...
    StatusEnvelope:
      type: object
      required: [schema, ok, data]
      properties: {schema: {type: string}, ok: {type: boolean}, data: {type: object}}
    EmbeddingBackfillEnvelope:
      type: object
      required: [schema, ok, data]
      properties:
        schema: {type: string}
        ok: {type: boolean}
        data:
          type: object
          required: [job_id, workspace_id, backlog, status]
          properties:
            job_id: {type: string}
            workspace_id: {type: string}
            mode: {type: string, enum: [missing, all]}
            queued: {type: integer}
            backlog: {type: integer}
            status: {type: string, enum: [catching_up, completed]}
    SearchPageEnvelope:
      type: object
      properties: {schema: {type: string}, ok: {type: boolean}, data: {$ref: '#/components/schemas/Page'}}
...