# server/http.go
package server

import (
	"crypto/subtle"
	"encoding/base64"
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"sort"
	"strconv"
	"strings"
	"sync"
	"time"

	"github.com/vessica-labs/vessica-knowledge-server/knowledge"
)

type Server struct {
	Service          *knowledge.Service
	Token            string
	ExportToken      string
	DefaultWorkspace string
	rateMu           sync.Mutex
	rate             map[string]rateWindow
	imports          map[string]map[string]any
}
type rateWindow struct {
	Started time.Time
	Count   int
}
type envelope struct {
	Schema string           `json:"schema"`
	OK     bool             `json:"ok"`
	Data   any              `json:"data,omitempty"`
	Error  *knowledge.Error `json:"error,omitempty"`
}

func (s *Server) Handler() http.Handler {
	m := http.NewServeMux()
	m.HandleFunc("GET /healthz", s.health)
	m.HandleFunc("GET /readyz", s.ready)
	m.HandleFunc("GET /v1/status", s.auth(s.status))
	m.HandleFunc("POST /admin/v1/embeddings/backfill", s.authWith(s.exportCredential(), s.embeddingBackfill))
	m.HandleFunc("GET /admin/v1/embeddings/backfill/{job_id}", s.authWith(s.exportCredential(), s.embeddingBackfillStatus))
	m.HandleFunc("GET /v1/search", s.auth(s.search))
	m.HandleFunc("POST /v1/context", s.auth(s.context))
	m.HandleFunc("POST /v1/memories:retrieve", s.auth(s.retrieveMemories))
	m.HandleFunc("POST /v1/scopes", s.auth(s.createScope))
	m.HandleFunc("GET /v1/scopes/{id}", s.auth(s.getScope))
	m.HandleFunc("POST /v1/entities", s.auth(s.createEntity))
	m.HandleFunc("GET /v1/entities", s.auth(s.listEntities))
	m.HandleFunc("GET /v1/entities/{id}", s.auth(s.getEntity))
	m.HandleFunc("GET /v1/entities:resolve", s.auth(s.resolveEntities))
	m.HandleFunc("POST /v1/artifacts", s.auth(s.createArtifact))
	m.HandleFunc("GET /v1/artifacts/{id}", s.auth(s.getArtifact))
	m.HandleFunc("GET /v1/artifacts/{id}/versions", s.auth(s.artifactVersions))
	m.HandleFunc("GET /v1/artifacts", s.auth(s.listArtifacts))
	m.HandleFunc("POST /v1/artifacts/{id}/versions", s.auth(s.versionArtifact))
	m.HandleFunc("POST /v1/artifacts/{id}/activate", s.auth(s.artifactStatus("active")))
	m.HandleFunc("POST /v1/artifacts/{id}/supersede", s.auth(s.artifactStatus("superseded")))
	m.HandleFunc("POST /v1/memories", s.auth(s.createMemory))
	m.HandleFunc("GET /v1/memories/{id}", s.auth(s.getMemory))
	m.HandleFunc("GET /v1/memories/{id}/versions", s.auth(s.memoryVersions))
	m.HandleFunc("GET /v1/memories", s.auth(s.searchMemories))
	m.HandleFunc("POST /v1/memories/{id}/versions", s.auth(s.versionMemory))
	m.HandleFunc("POST /v1/memories/{id}/supersede", s.auth(s.memoryState("superseded")))
	m.HandleFunc("POST /v1/memories/{id}/archive", s.auth(s.memoryState("archived")))
	m.HandleFunc("POST /v1/relationships", s.auth(s.createRelationship))
	m.HandleFunc("GET /v1/relationships", s.auth(s.listRelationships))
	m.HandleFunc("POST /v1/workflow-events", s.auth(s.workflow))
	m.HandleFunc("POST /v1/exports", s.authWith(s.exportCredential(), s.export))
	m.HandleFunc("POST /v1/imports", s.authWith(s.exportCredential(), s.importSnapshot))
	m.HandleFunc("GET /v1/imports/{id}", s.authWith(s.exportCredential(), s.importStatus))
	// Go's ServeMux does not support a static suffix after a path wildcard.
	// Preserve the public colon-action API by rewriting it to internal routes.
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		for from, to := range map[string]string{":version": "/versions", ":activate": "/activate", ":supersede": "/supersede", ":archive": "/archive"} {
			if strings.HasSuffix(r.URL.Path, from) {
				r.URL.Path = strings.TrimSuffix(r.URL.Path, from) + to
				break
			}
		}
		m.ServeHTTP(w, r)
	})
}

func (s *Server) embeddingBackfill(w http.ResponseWriter, r *http.Request) {
	var input struct {
		WorkspaceID string `json:"workspace_id"`
		Mode        string `json:"mode"`
	}
	if !decode(w, r, &input) {
		return
	}
	if input.WorkspaceID == "" {
		input.WorkspaceID = s.DefaultWorkspace
	}
	if s.DefaultWorkspace != "" && input.WorkspaceID != s.DefaultWorkspace {
		failure(w, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403})
		return
	}
	queued, err := s.Service.Store.QueueEmbeddings(r.Context(), input.WorkspaceID, input.Mode)
	if err != nil {
		failure(w, err)
		return
	}
	backlog, err := s.Service.Store.EmbeddingBacklog(r.Context(), input.WorkspaceID)
	if err != nil {
		failure(w, err)
		return
	}
	mode := input.Mode
	if mode == "" {
		mode = "missing"
	}
	respond(w, knowledge.EmbeddingBackfill{JobID: "emb_" + strconv.FormatInt(time.Now().UTC().UnixNano(), 36), WorkspaceID: input.WorkspaceID, Mode: mode, Queued: queued, Backlog: backlog, Status: backfillStatus(backlog)}, nil, http.StatusAccepted)
}

func (s *Server) embeddingBackfillStatus(w http.ResponseWriter, r *http.Request) {
	ws := workspace(r, s.DefaultWorkspace)
	backlog, err := s.Service.Store.EmbeddingBacklog(r.Context(), ws)
	if err != nil {
		failure(w, err)
		return
	}
	respond(w, knowledge.EmbeddingBackfill{JobID: r.PathValue("job_id"), WorkspaceID: ws, Backlog: backlog, Status: backfillStatus(backlog)}, nil, http.StatusOK)
}

func backfillStatus(backlog int) string {
	if backlog > 0 {
		return "catching_up"
	}
	return "completed"
}
func (s *Server) status(w http.ResponseWriter, r *http.Request) {
	backlog, err := s.Service.Store.EmbeddingBacklog(r.Context(), workspace(r, s.DefaultWorkspace))
	if err != nil {
		failure(w, err)
		return
	}
	result := knowledge.Status{Schema: knowledge.APIVersion, RetrievalMode: "lexical", EmbeddingState: "not_configured", EmbeddingBacklog: backlog, IndexFresh: backlog == 0}
	if s.Service.Embedder != nil {
		result.RetrievalMode = "semantic_hybrid"
		result.EmbeddingState = "configured"
		result.EmbeddingModel = s.Service.Embedder.Model()
	}
	if s.Service.Reranker != nil {
		result.RerankEnabled = true
		result.RerankModel = s.Service.Reranker.Model()
	}
	respond(w, result, nil, http.StatusOK)
}

func cursorOffset(raw string) int {
	decoded, err := base64.RawURLEncoding.DecodeString(raw)
	if err != nil {
		return 0
	}
	n, _ := strconv.Atoi(string(decoded))
	if n < 0 {
		return 0
	}
	return n
}
func pageSize(r *http.Request) int {
	n, _ := strconv.Atoi(r.URL.Query().Get("limit"))
	if n <= 0 {
		return 50
	}
	if n > 200 {
		return 200
	}
	return n
}
func paginate[T any](items []T, r *http.Request) knowledge.Page[T] {
	offset, limit := cursorOffset(r.URL.Query().Get("cursor")), pageSize(r)
	if offset > len(items) {
		offset = len(items)
	}
	end := offset + limit
	if end > len(items) {
		end = len(items)
	}
	result := knowledge.Page[T]{Items: items[offset:end]}
	if end < len(items) {
		result.NextCursor = base64.RawURLEncoding.EncodeToString([]byte(strconv.Itoa(end)))
	}
	return result
}

func (s *Server) search(w http.ResponseWriter, r *http.Request) {
	ws, q, scopes := workspace(r, s.DefaultWorkspace), strings.TrimSpace(r.URL.Query().Get("q")), r.URL.Query()["scope"]
	entities, err := s.Service.Store.ResolveEntities(r.Context(), ws, scopes, q)
	if err != nil {
		failure(w, err)
		return
	}
	artifacts, err := s.Service.Store.ListArtifacts(r.Context(), ws, scopes, nil)
	if err != nil {
		failure(w, err)
		return
	}
	memories, err := s.Service.Store.SearchMemories(r.Context(), ws, scopes, q, 500)
	if err != nil {
		failure(w, err)
		return
	}
	want := strings.TrimSpace(r.URL.Query().Get("object_type"))
	var results []knowledge.SearchResult
	if want == "" || want == "entity" {
		for _, v := range entities {
			results = append(results, knowledge.SearchResult{ObjectType: "entity", ID: v.ID, ScopeID: v.ScopeID, Subtype: v.Type, Title: v.DisplayName, State: v.State, UpdatedAt: v.UpdatedAt})
		}
	}
	if want == "" || want == "artifact" {
		for _, v := range artifacts {
			if q == "" || strings.Contains(strings.ToLower(v.Title+" "+v.Content), strings.ToLower(q)) {
				results = append(results, knowledge.SearchResult{ObjectType: "artifact", ID: v.ID, ScopeID: v.ScopeID, Subtype: v.Type, Title: v.Title, Summary: truncate(v.Content, 240), State: v.Status, UpdatedAt: v.UpdatedAt})
			}
		}
	}
	if want == "" || want == "memory" {
		for _, v := range memories {
			results = append(results, knowledge.SearchResult{ObjectType: "memory", ID: v.ID, ScopeID: v.ScopeID, Subtype: v.Type, Title: v.Title, Summary: truncate(v.Content, 240), State: v.State, UpdatedAt: v.UpdatedAt})
		}
	}
	sort.SliceStable(results, func(i, j int) bool { return results[i].UpdatedAt.After(results[j].UpdatedAt) })
	respond(w, paginate(results, r), nil, http.StatusOK)
}
func truncate(v string, n int) string {
	if len(v) <= n {
		return v
	}
	return v[:n] + "…"
}
func (s *Server) health(w http.ResponseWriter, r *http.Request) {
	write(w, 200, map[string]any{"service": "vessica-knowledge-server", "ok": true})
}
func (s *Server) ready(w http.ResponseWriter, r *http.Request) {
	if s.Service == nil || s.Service.Store.Ping(r.Context()) != nil {
		write(w, 503, map[string]any{"ok": false, "api": "ready", "database": "unavailable", "migrations": "ready", "embedding_worker": "unknown", "index": "unknown"})
		return
	}
	embedding := "not_configured"
	index := "ready"
	if s.Service.Embedder != nil {
		embedding = "ready"
		if backlog, _ := s.Service.Store.EmbeddingBacklog(r.Context(), s.DefaultWorkspace); backlog > 0 {
			index = "catching_up"
		}
	}
	write(w, 200, map[string]any{"ok": true, "api": "ready", "database": "ready", "migrations": "ready", "embedding_worker": embedding, "index": index})
}
func (s *Server) auth(next http.HandlerFunc) http.HandlerFunc {
	return s.authWith(s.Token, next)
}
func (s *Server) exportCredential() string {
	if s.ExportToken != "" {
		return s.ExportToken
	}
	return s.Token
}
func (s *Server) authWith(required string, next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if required != "" && !token(r.Header.Get("Authorization"), required) {
			failure(w, &knowledge.Error{Code: "unauthorized", Message: "invalid bearer token", Status: 401})
			return
		}
		if !s.allow(r.Header.Get("Authorization")) {
			failure(w, &knowledge.Error{Code: "rate_limited", Message: "request rate limit exceeded", Status: 429})
			return
		}
		next(w, r)
	}
}
func (s *Server) allow(key string) bool {
	s.rateMu.Lock()
	defer s.rateMu.Unlock()
	if s.rate == nil {
		s.rate = map[string]rateWindow{}
	}
	now := time.Now()
	v := s.rate[key]
	if v.Started.IsZero() || now.Sub(v.Started) >= time.Minute {
		v = rateWindow{Started: now}
	}
	v.Count++
	s.rate[key] = v
	return v.Count <= 600
}
func token(h, want string) bool {
	got := strings.TrimSpace(strings.TrimPrefix(h, "Bearer "))
	if len(got) != len(want) {
		return false
	}
	return subtle.ConstantTimeCompare([]byte(got), []byte(want)) == 1
}
func decode(w http.ResponseWriter, r *http.Request, v any) bool {
	d := json.NewDecoder(http.MaxBytesReader(w, r.Body, 64<<20))
	d.DisallowUnknownFields()
	if err := d.Decode(v); err != nil {
		failure(w, &knowledge.Error{Code: "invalid_json", Message: err.Error(), Status: 400})
		return false
	}
	return true
}
func (s *Server) opts(r *http.Request, workspace string) (knowledge.WriteOptions, error) {
	if workspace == "" {
		workspace = s.DefaultWorkspace
	}
	if s.DefaultWorkspace != "" && workspace != s.DefaultWorkspace {
		return knowledge.WriteOptions{}, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403}
	}
	o := knowledge.WriteOptions{WorkspaceID: workspace, IdempotencyKey: r.Header.Get("Idempotency-Key"), Actor: knowledge.Actor{ID: r.Header.Get("X-Actor-ID"), Type: r.Header.Get("X-Actor-Type")}, Provenance: knowledge.Provenance{Source: r.Header.Get("X-Provenance-Source"), SourceID: r.Header.Get("X-Provenance-ID")}}
	if o.Actor.ID == "" || o.Actor.Type == "" {
		return o, knowledge.Invalid("X-Actor-ID and X-Actor-Type headers required")
	}
	if o.Provenance.Source == "" {
		return o, knowledge.Invalid("X-Provenance-Source header required")
	}
	if o.IdempotencyKey == "" {
		return o, knowledge.Invalid("Idempotency-Key header required")
	}
	return o, nil
}
func workspace(r *http.Request, def string) string {
	if def != "" {
		return def
	}
	if v := r.URL.Query().Get("workspace_id"); v != "" {
		return v
	}
	return def
}
func (s *Server) context(w http.ResponseWriter, r *http.Request) {
	var v knowledge.ContextRequest
	if !decode(w, r, &v) {
		return
	}
	if v.WorkspaceID == "" {
		v.WorkspaceID = s.DefaultWorkspace
	}
	if s.DefaultWorkspace != "" && v.WorkspaceID != s.DefaultWorkspace {
		failure(w, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403})
		return
	}
	got, err := s.Service.Context(r.Context(), v)
	respond(w, got, err, 200)
}

func (s *Server) retrieveMemories(w http.ResponseWriter, r *http.Request) {
	var request knowledge.MemoryRetrievalRequest
	if !decode(w, r, &request) {
		return
	}
	if request.WorkspaceID == "" {
		request.WorkspaceID = s.DefaultWorkspace
	}
	if s.DefaultWorkspace != "" && request.WorkspaceID != s.DefaultWorkspace {
		failure(w, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403})
		return
	}
	if strings.TrimSpace(request.Query) == "" || len(request.ScopeIDs) == 0 {
		respond(w, nil, knowledge.Invalid("query and at least one scope are required"), http.StatusBadRequest)
		return
	}
	result, err := s.Service.RetrieveMemories(r.Context(), request)
	respond(w, result, err, http.StatusOK)
}
func (s *Server) createScope(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Scope
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.CreateScope(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) getScope(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.GetScope(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"))
	respond(w, got, e, 200)
}
func (s *Server) createEntity(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Entity
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.CreateEntity(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) getEntity(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.GetEntity(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"))
	respond(w, got, e, 200)
}
func (s *Server) listEntities(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.ListEntities(r.Context(), workspace(r, s.DefaultWorkspace), r.URL.Query()["scope"], r.URL.Query().Get("type"), r.URL.Query().Get("state"))
	if e != nil {
		failure(w, e)
		return
	}
	respond(w, paginate(got, r), nil, http.StatusOK)
}
func (s *Server) resolveEntities(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.ResolveEntities(r.Context(), workspace(r, s.DefaultWorkspace), r.URL.Query()["scope"], r.URL.Query().Get("q"))
	respond(w, got, e, 200)
}
func (s *Server) createArtifact(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Artifact
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.CreateArtifact(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) versionArtifact(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Artifact
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.VersionArtifact(r.Context(), o, r.PathValue("id"), v)
	respond(w, got, e, 200)
}
func (s *Server) artifactStatus(status string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var body struct {
			WorkspaceID string `json:"workspace_id"`
		}
		if !decode(w, r, &body) {
			return
		}
		o, e := s.opts(r, body.WorkspaceID)
		if e != nil {
			failure(w, e)
			return
		}
		got, e := s.Service.SetArtifactStatus(r.Context(), o, r.PathValue("id"), status)
		respond(w, got, e, 200)
	}
}
func (s *Server) getArtifact(w http.ResponseWriter, r *http.Request) {
	n, _ := strconv.Atoi(r.URL.Query().Get("version"))
	got, e := s.Service.Store.GetArtifact(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"), n)
	respond(w, got, e, 200)
}
func (s *Server) artifactVersions(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.ListArtifactVersions(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"))
	if e != nil {
		failure(w, e)
		return
	}
	respond(w, paginate(got, r), nil, http.StatusOK)
}
func (s *Server) listArtifacts(w http.ResponseWriter, r *http.Request) {
	sel := knowledge.ArtifactSelector{Type: r.URL.Query().Get("type"), Status: r.URL.Query().Get("status")}
	got, e := s.Service.Store.ListArtifacts(r.Context(), workspace(r, s.DefaultWorkspace), r.URL.Query()["scope"], []knowledge.ArtifactSelector{sel})
	respond(w, got, e, 200)
}
func (s *Server) createMemory(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Memory
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.CreateMemory(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) versionMemory(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Memory
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.VersionMemory(r.Context(), o, r.PathValue("id"), v)
	respond(w, got, e, 200)
}
func (s *Server) memoryState(state string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var body struct {
			WorkspaceID string `json:"workspace_id"`
		}
		if !decode(w, r, &body) {
			return
		}
		o, e := s.opts(r, body.WorkspaceID)
		if e != nil {
			failure(w, e)
			return
		}
		got, e := s.Service.SetMemoryState(r.Context(), o, r.PathValue("id"), state)
		respond(w, got, e, 200)
	}
}
func (s *Server) getMemory(w http.ResponseWriter, r *http.Request) {
	n, _ := strconv.Atoi(r.URL.Query().Get("version"))
	got, e := s.Service.Store.GetMemory(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"), n)
	respond(w, got, e, 200)
}
func (s *Server) memoryVersions(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.ListMemoryVersions(r.Context(), workspace(r, s.DefaultWorkspace), r.PathValue("id"))
	if e != nil {
		failure(w, e)
		return
	}
	respond(w, paginate(got, r), nil, http.StatusOK)
}
func (s *Server) searchMemories(w http.ResponseWriter, r *http.Request) {
	got, e := s.Service.Store.SearchMemories(r.Context(), workspace(r, s.DefaultWorkspace), r.URL.Query()["scope"], r.URL.Query().Get("q"), 100)
	respond(w, got, e, 200)
}
func (s *Server) createRelationship(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Relationship
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.CreateRelationship(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) listRelationships(w http.ResponseWriter, r *http.Request) {
	objectID := r.URL.Query().Get("object_id")
	if objectID == "" {
		failure(w, knowledge.Invalid("object_id required"))
		return
	}
	got, e := s.Service.Store.ListRelationships(r.Context(), workspace(r, s.DefaultWorkspace), objectID)
	if e != nil {
		failure(w, e)
		return
	}
	respond(w, paginate(got, r), nil, http.StatusOK)
}
func (s *Server) workflow(w http.ResponseWriter, r *http.Request) {
	var v knowledge.WorkflowEvent
	if !decode(w, r, &v) {
		return
	}
	o, e := s.opts(r, v.WorkspaceID)
	if e != nil {
		failure(w, e)
		return
	}
	got, e := s.Service.IngestWorkflowEvent(r.Context(), o, v)
	respond(w, got, e, 201)
}
func (s *Server) export(w http.ResponseWriter, r *http.Request) {
	var body struct {
		WorkspaceID string `json:"workspace_id"`
	}
	if !decode(w, r, &body) {
		return
	}
	if s.DefaultWorkspace != "" && body.WorkspaceID != s.DefaultWorkspace {
		failure(w, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403})
		return
	}
	got, e := s.Service.Store.Export(r.Context(), body.WorkspaceID)
	respond(w, got, e, 200)
}
func (s *Server) importSnapshot(w http.ResponseWriter, r *http.Request) {
	var v knowledge.Snapshot
	if !decode(w, r, &v) {
		return
	}
	if s.DefaultWorkspace != "" && v.WorkspaceID != s.DefaultWorkspace {
		failure(w, &knowledge.Error{Code: "forbidden_workspace", Message: "token is not scoped to this workspace", Status: 403})
		return
	}
	e := s.Service.ImportSnapshot(r.Context(), v)
	id := "import-" + v.Checksum
	status := map[string]any{"id": id, "status": map[bool]string{true: "completed", false: "failed"}[e == nil], "imported": e == nil, "counts": v.Counts, "checksum": v.Checksum}
	s.rateMu.Lock()
	if s.imports == nil {
		s.imports = map[string]map[string]any{}
	}
	s.imports[id] = status
	s.rateMu.Unlock()
	respond(w, status, e, 201)
}
func (s *Server) importStatus(w http.ResponseWriter, r *http.Request) {
	s.rateMu.Lock()
	v, ok := s.imports[r.PathValue("id")]
	s.rateMu.Unlock()
	if !ok {
		failure(w, knowledge.NotFound("import not found"))
		return
	}
	respond(w, v, nil, 200)
}
func respond(w http.ResponseWriter, data any, err error, status int) {
	if err != nil {
		failure(w, err)
		return
	}
	writeEnvelope(w, status, envelope{Schema: knowledge.APIVersion, OK: true, Data: data})
}
func failure(w http.ResponseWriter, err error) {
	var ke *knowledge.Error
	if !errors.As(err, &ke) {
		ke = &knowledge.Error{Code: "internal", Message: err.Error(), Status: 500}
	}
	writeEnvelope(w, ke.Status, envelope{Schema: knowledge.APIVersion, OK: false, Error: ke})
}
func write(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}
func writeEnvelope(w http.ResponseWriter, status int, v envelope) { write(w, status, v) }

var _ = io.EOF